Fix 49 tests: strings, arrays, sscanf, html entities, similar_text
Strings (+36):
- Rewrite wordwrap, str_replace (array support), html_entity_decode
- Implement similar_text with longest-common-substring algorithm
- Fix addcslashes, strip_tags (quoted attrs), strncmp ValueError
- Fix sprintf, crc32 unsigned return, substr_compare
- Remove 15 duplicate function registrations in misc.rs

Arrays (+12):
- Implement array_column, array_all/any with callbacks
- Fix array_combine validation, array_diff/diff_key/diff_assoc 3+ arrays
- Fix array_change_key_case TypeError, array_count_values warnings

General:
- Implement sscanf (%d, %s, %f, %x, %o, %c, %n, %[])
- Fix htmlspecialchars ENT_COMPAT/QUOTES/NOQUOTES flags
- Implement htmlspecialchars_decode
- SendRef for str_replace &$count, similar_text &$percent

+++ crates/goro-ext-hash/src/lib.rs
@@ -31,8 +31,8 @@ fn crc32_fn(_vm: &mut Vm, args: &[Value]) -> Result<Value, VmError> {
         }
     }
     crc ^= 0xFFFFFFFF;
-    // PHP returns signed integer
-    Ok(Value::Long(crc as i32 as i64))
+    // PHP 64-bit returns unsigned 32-bit value as positive integer
+    Ok(Value::Long(crc as i64))
 }
 fn md5_fn(_vm: &mut Vm, args: &[Value]) -> Result<Value, VmError> {
     let data = args.first().unwrap_or(&Value::Null).to_php_string();
